requirements.txt
pyenchant
more_itertools